Give more meaningful name to multiple database columns(#99)
In preparation of the SSDT, it was decided to give better name to avoid confusion.

* Rename tables SlotNumberType to Attendance and TrainingSlot to TrainingAttendance

* Rename tables TrainingIdentity to VatExemptionClaim and TrainingType to VatExemptionType

* Rename Training's property Status to StatusType

* Rename Enumeration TrainingTopic to Topic

* Rename TrainingCategory to TrainingTopic

* Rename TrainingDetail to TrainingLocalizedDetails

* Rename PersonalSocialNetwork to TrainerSocialNetwork

* Rename enumeration TrainingTargetAudience to TargetAudienceType

* Rename TrainingTarget to TrainingTargetAudience

* Rename unmatching names in dapper queries

* Rename foreign key trainer social network

+++ src/Showcase/src/Smart.FA.Catalog.Showcase.Domain/Models/TrainingList.cs
@@ -2,6 +2,7 @@
 #nullable enable
 
 namespace Smart.FA.Catalog.Showcase.Domain.Models;
+
 public partial class TrainingList
 {
     public int Id { get; set; }
@@ -10,6 +11,6 @@ public partial class TrainingList
     public string TrainerFirstName { get; set; } = null!;
     public string TrainerLastName { get; set; } = null!;
     public int Topic { get; set; }
-    public short Status { get; set; }
+    public int Status { get; set; }
     public int TrainerId { get; set; }
 }
